requirements.txt
scikit-image==0.19.2
Pillow==9.0.1
tqdm==4.63.1
numpy==1.21.5
torch==1.10.0
torchvision==0.11.0
pytorch-lightning==1.6.0
